adding catboost

+++ crosspredict/crossval/_crossval.py
@@ -244,7 +244,7 @@ class CrossModelFabric(ABC):
         model = self.models[0]
         explainer = shap.TreeExplainer(model)
         df_sample = test[model.feature_name()].sample(
-            n=n_samples, random_state=0, replace=True).astype(float)
+            n=n_samples, random_state=0, replace=True)
         if self.params['metric']=='auc':
             shap_values = explainer.shap_values(df_sample)[1]
         else:
